allert

+++ components/FavoriteComponent.js
@@ -1,12 +1,13 @@
+
 import React, { Component } from 'react';
-import { View, FlatList, Text } from 'react-native';
+
+import { Alert, FlatList, Text, View } from 'react-native';
 import { ListItem } from 'react-native-elements';
 import { connect } from 'react-redux';
 import { baseUrl } from '../shared/baseUrl';
 import { Loading } from './LoadingComponent';
 import Swipeout from 'react-native-swipeout';
 import { deleteFavorite } from '../redux/ActionCreators';
-
 const mapStateToProps = state => {
     return {
         dishes: state.dishes,
@@ -15,49 +16,54 @@ const mapStateToProps = state => {
 };
 const mapDispatchToProps = dispatch => ({
     deleteFavorite: (dishId) => dispatch(deleteFavorite(dishId))
-})
-
+});
 class Favorites extends Component {
-
     static navigationOptions = {
         title: 'My Favorites'
     }
-
     render() {
         const { navigate } = this.props.navigation;
-        const rightButton = [
-            {
-                text: 'Delete', 
-                type: 'delete',
-                onPress: () => this.props.deleteFavorite(item.id)
-            }
-        ];
-
-        return (
-            <Swipeout right={rightButton} autoClose={true}>
-                <ListItem
-                    key={index}
-                    title={item.name}
-                    subtitle={item.description}
-                    hideChevron={true}
-                    onPress={() => navigate('Dishdetail', { dishId: item.id })}
-                    leftAvatar={{ source: {uri: baseUrl + item.image}}}
-                    />
-            </Swipeout>
-        );
         const renderMenuItem = ({ item, index }) => {
+            const rightButton = [
+                {
+                    text: 'Delete',
+                    type: 'delete',
+                  
+                    onPress: () => {
+                        Alert.alert(
+                            'Delete Favorite?',
+                            'Are you sure you wish to delete the favorite dish ' + item.name + '?',
+                            [
+                                { 
+                                    text: 'Cancel', 
+                                    onPress: () => console.log(item.name + 'Not Deleted'),
+                                    style: 'cancel',
+                                },
+                                { 
+                                    text: 'OK', 
+                                    onPress: () => this.props.deleteFavorite(item.id),
+                                    style: 'destructive'
+                                }
+                            ],
+                            { cancelable: false }
+                        ); 
+                    }
+                }
+            ];
+
             return(
-                <ListItem
-                    key={index}
-                    title={item.name}
-                    subtitle={item.description}
-                    hideChevron={true}
-                    onPress={() => navigate('Dishdetail', { dishId: item.id })}
-                    leftAvatar={{ source: { uri: baseUrl + item.image }}}
-                />
+                <Swipeout right={rightButton} autoClose={true}>
+                    <ListItem
+                        key={index}
+                        title={item.name}
+                        subtitle={item.description}
+                        hideChevron={true}
+                        onPress={() => navigate('Dishdetail', { dishId: item.id })}
+                        leftAvatar={{ source: { uri: baseUrl + item.image }}}
+                    />
+                </Swipeout>
             );
         };
-
         if (this.props.dishes.isLoading) {
             return(
                 <Loading />
@@ -81,5 +87,4 @@ class Favorites extends Component {
         }
     }
 };
-
 export default connect(mapStateToProps, mapDispatchToProps)(Favorites);
